Refactor project folder structure
