+++ src/KiBoards/Models/Settings/KibanaSettingsChanges.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace KiBoards.Services
+namespace KiBoards.Models.Settings
 {
     class KibanaSettingsChanges
     {
