+++ src/KiBoards/Models/Settings/KibanaSettingsRequest.cs
@@ -1,5 +1,5 @@
-﻿namespace KiBoards.Services
-{ 
+﻿namespace KiBoards.Models.Settings
+{
     class KibanaSettingsRequest
     {
         public KibanaSettingsChanges Changes { get; set; }
