+++ src/KiBoards/Models/Status/KibanaStatusOverall.cs
@@ -1,4 +1,4 @@
-﻿namespace KiBoards.Services
+﻿namespace KiBoards.Models.Status
 {
     class KibanaStatusOverall
     {
