+++ src/KiBoards/Models/Status/KibanaStatusResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace KiBoards.Services
+﻿namespace KiBoards.Models.Status
 {
     class KibanaStatusResponse
     {
