+++ src/KiBoards/Models/Status/KibanaVersion.cs
@@ -1,4 +1,4 @@
-﻿namespace KiBoards.Services
+﻿namespace KiBoards.Models.Status
 {
     class KibanaVersion
     {
