+++ src/KiBoards/Startup.cs
@@ -18,10 +18,10 @@ namespace KiBoards
             var task = Task.Factory.StartNew(async () =>
             {
                 var httpClient = new HttpClient();
-                var kibanaHost = Environment.GetEnvironmentVariable("KIB_KIBANA_HOST") ?? "http://localhost:5601";
-                var kibanaClient = new KibanaClient(httpClient, new Uri(kibanaHost));
+                var kibanaUri = new Uri(Environment.GetEnvironmentVariable("KIB_KIBANA_HOST") ?? "http://localhost:5601");
+                var kibanaClient = new KiBoardsKibanaClient(kibanaUri, httpClient);
 
-                WriteMessage(messageSink, $"Waiting for Kibana {kibanaHost}");
+                WriteMessage(messageSink, $"Waiting for Kibana {kibanaUri}");
 
                 await kibanaClient.WaitForKibanaAsync(CancellationToken.None);
                 await kibanaClient.SetDarkModeAsync(true, CancellationToken.None);
